use better interceptor
