--- airbyte-cdk/bulk/core/base/src/testFixtures/resources/META-INF.services/org.junit.jupiter.api.extension.Extension
@@ -1,1 +1,0 @@
-io.airbyte.cdk.extensions.LoggingInvocationInterceptor
