+++ airbyte-cdk/bulk/core/base/src/testFixtures/resources/META-INF/services/org.junit.jupiter.api.extension.Extension
@@ -1,0 +1,1 @@
+io.airbyte.cdk.extensions.JunitTimeoutInvocationInterceptor
